Extend LazyMobXTable to have resizable columns
This changes are supposed to add a new boolean prop to the table column so that whenever it's set to true it will make the column resizable. @ts-ignore was added to avoid failed checking of untyped JS lib. These changes also include making the Sample column in both CNA and Mutation tables resizable.